Rename file_op.py to operations.py
