+++ src/pyscaffold/extensions/cirrus.py
@@ -3,7 +3,7 @@
 import argparse
 
 from ..api import Extension, helpers
-from ..file_op import no_overwrite
+from ..operations import no_overwrite
 from ..templates import get_template
 from .pre_commit import PreCommit
 from .tox import Tox
